Remove RunPod DINO runtime
