+++ .env.example
@@ -20,14 +20,9 @@ MODAL_DINO_FUNCTION=infer
 MODAL_ENVIRONMENT=main
 MODAL_PROFILE=
 
-# Hosted GroundingDINO + DINOv2 endpoint. When set, detection and embeddings are API calls and
-# torch is never loaded by the pipeline process.
-RUNPOD_DINO_ENDPOINT=
-RUNPOD_API_KEY=
-#
 # B. …or a local GPU TRELLIS install (Linux + NVIDIA only) — that env's python interpreter.
 TRELLIS_PYTHON=
 
 # --- isolated detection runtime ---
-# Python on the machine/environment where GroundingDINO is allowed to run.
+# Optional alternative to Modal DINO: an explicit local GroundingDINO environment.
 GROUNDING_DINO_PYTHON=
