+++ ARCHITECTURE.md
@@ -26,11 +26,10 @@ src/litereality_agent/
 ├── room_format/         Room.py manifest, compilation, rendering, export, and validation
 ├── models/              canonical inference and model-specific service adapters
 │   └── llm/             provider integrations (`openai/` and `claude/`)
-└── runtimes/            execution transports such as Modal and RunPod
+└── runtimes/            hosted execution transports such as Modal
 
 scripts/                  standalone capture and publishing utilities
 deploy/modal/             hosted model wrappers; never imported by application code
-deploy/runpod/            remaining DINO container packaging
 ```
 
 There are no `services`, `adapters`, `shared`, or nested `pipeline/stages` layers. The directory
@@ -86,7 +85,8 @@ process environment > .env > models.env > typed defaults
 Heavy inference is isolated from the main environment. A model package owns one inference path
 regardless of where it executes. A local adapter can run it in a separately configured process;
 a hosted adapter sends its request contract through `runtimes/`. Modal wraps canonical TRELLIS,
-GroundingDINO, and DINOv2 inference under `deploy/modal`; RunPod remains an optional DINO fallback.
+GroundingDINO, and DINOv2 inference under `deploy/modal`; DINO also supports an explicitly selected
+local model environment.
 Deployment files stay outside application source. The CLI and unit tests do not start DINO,
 TRELLIS, Blender, or paid model calls.
