--- deploy/runpod/grounding_dino/requirements.txt
@@ -1,4 +1,0 @@
-runpod~=1.7.0
-transformers>=4.51,<6
-pillow>=10
-numpy<2
