+++ src/litereality_agent/models/registry.py
@@ -43,15 +43,6 @@ def detection_from_settings(settings: LiteRealitySettings | None = None):
             model_id=settings.dino_model,
             embed_model_id=settings.dino_embed_model,
         )
-    if settings.runpod_dino_endpoint:
-        from litereality_agent.models.grounding_dino.runpod import RunPodDinoService
-
-        return RunPodDinoService(
-            api_key=settings.runpod_api_key.get_secret_value() if settings.runpod_api_key else None,
-            endpoint_id=settings.runpod_dino_endpoint,
-            model_id=settings.dino_model,
-            embed_model_id=settings.dino_embed_model,
-        )
     if settings.dino_python:
         from litereality_agent.models.grounding_dino.service import DinoSubprocessService
